Use a pool for APN connections
The connections to APN are the only connections in pling that are
steadily held open (Apple requires this).

Without a connection pool there might be race-conditions when using
pling in a multi-threaded context.

This adds a connection pool so that only one thread at a time can use an
open connection to APN.