Make category attribute available for APN messages
This extends `Pling::Message` to have a `category` property. When
setting the `category` it always gets converted to a string. If not set,
it's `nil`.

`APN::Gateway` then checks whether `category` is set (truthy) and adds
it to the `aps` hash of the message.

According to the Apple documentation the category has to correspond to a
custom value defined in the receiving application and should be a
string.

    https://developer.apple.com/library/ios/documentation/NetworkingInternet/Conceptual/RemoteNotificationsPG/Chapters/ApplePushService.html#//apple_ref/doc/uid/TP40008194-CH100-SW1

If the category is not set, we do not try to send it to Apple, to keep
the default behaviour.