Fix nil-error when APN returns no tokens
This error was caught in the wild and happens when APN doesn't return any data,
but just "data" that results in `length` being `nil`.

This adds a test case that reproduces the error and fixes it by stopping
the fetching of data as soon as `nil` is encountered.

Also: an unused local variable is removed by replacing it with the `_`.